refactor(config): the fifth ratchet raise became a split, and a leaf is why it works
`config/types.ts` is the config contract for every subsystem -- spaces, networks, media, embedding, auth, sync.
It took FOUR god-file ratchet raises in two days, each individually correct and each one line of honest type.
The gate's own comment said what to do about a fifth: "If it wants a fifth, do the split first instead -- that
is the signal, not the number." Q-2 wants a fifth, for `suppressEmbeddings` on `TypeSchema`. So: the split.

The knowledge-schema vocabulary moves to `config/types-knowledge.ts` -- merge functions, `PropertySchema`,
`TypeSchema`, `ValidationMode`, `KnowledgeType`, and `SpaceMeta`, which is the thing that holds the schemas.
Re-exported from `types.ts`, so NO importer changes and no behaviour change. Ratchet lowered 677 -> 645.

The new file imports nothing, and that is the load-bearing property rather than a tidiness preference.

The first attempt at this split moved the NETWORK types out instead and re-exported them. It compiled both
moved files clean and took `types.ts` from 677 to 609. It was also wrong: `NetworkConfig` references
`SpaceMeta`, which belongs to spaces and stayed behind, so `types.ts` re-exported the new file while the new
file imported `SpaceMeta` back. A genuine module cycle. TypeScript resolved it by degrading `NetworkConfig` to
`any`, and `api/invite.ts` picked up three `TS7006` implicit-anys on its `members` callbacks -- a caller
silently losing its types while every file in the diff built clean. That is the same shape as the bug in #773:
a check that under-reports reads exactly like a pass. It was reverted rather than shipped.

A leaf cannot be half of a cycle. Same reasoning as `config/rights-shape.ts`, which exists for exactly this.

So the verification here is on the CALLERS, not on the moved files:

    npm --prefix server run build 2>&1 | grep -c "TS7006"    # 0

`SpaceMeta` needed more company than expected, and that is what makes the leaf the right unit: it depends on
`ValidationMode`, and through `typeSchemas` on `KnowledgeType` and `TypeSchema`, which in turn rests on
`PropertySchema` and the merge functions. That set is one domain -- the vocabulary a schema is written in --
so it moves as one thing. It is also where per-type schema fields now GROW, which is the actual unblocking:
Q-2's field lands in the leaf instead of raising this file a fifth time.

One gate failed on the move and was right to: `documented-interfaces-match-code.test.js` pins
`TypeSchema`/`PropertySchema` to a source path, and asserts it FOUND both declarations. A gate that resolved a
stale path to nothing and passed would be the vacuous kind. Paths updated, and the reason recorded inline.
